chore: Update composer.json to use "packages\\Razor\\Paystack\\Providers\\PaystackServiceProvider"
The `composer.json` file is modified to update the provider for the Paystack package. The previous provider "Razor\\Paystack\\Providers\\PaystackServiceProvider" is replaced with "packages\\Razor\\Paystack\\Providers\\PaystackServiceProvider". This change ensures that the correct provider is used for the Paystack integration in the project.

Note: The recent user commits and recent repository commits are not relevant for generating the commit message.